Add files via upload
This project is the Proverif code of the paper "A Secure and Efficient Three-Factor Authentication Protocol for IoT Environments", which is used to verify the security of the protocol proposed in the paper.

The following are the necessary explanations:

The file "SE3FA.pv" is the Proverif implementation of the proposed protocol under general assumptions, and the file "SE3FA(long_term_secret_exposed).pv" is the Proverif implementation of the proposed protocol in the case that adversary knows the long term secret information.

The file "result_of_SE3FA.txt" is a record of the result produced by running "SE3FA.pv". The file "result_of_SE3FA(long_term_secret_exposed).txt" is a record of the result produced by running "SE3FA(long_term_secret_exposed).pv".